requirements.txt
colorama==0.4.4
prompt-toolkit==1.0.14
pyfiglet==0.8.post1
Pygments==2.11.2
PyInquirer==1.0.3
python-cfonts==1.5.2
regex==2022.1.18
six==1.16.0
termcolor==1.1.0
wcwidth==0.2.5